Build the proper file structure
