+++ app.py
@@ -1,60 +1,51 @@
 import faster_than_requests as req
 import tkinter as tk
-from glom import glom
 import json
 
 
 def request_data():
-    xdarr = ['First Name', 'Last Name', 'Gender', 'Address', 'City', 'State', 'Country', 'Latitude', 'Longitude','Postal Code', 'Time Zone', 'Email', 'Birthday', 'Age', 'Telephone', 'Cell Phone']
     returnData = req.get2json_pretty("https://randomuser.me/api/")
-    returnData = json.loads(returnData)
+    returnData = dict(json.loads(returnData))
+    results = returnData['results'][0]
+    # Adding results into a array
+    rec(results)
+    # formatting & writing data into the disk
+    writedata(bindstr(arrkeys, arr))
 
-    # Adding JSON data into a array using generator
-    arr = []
-    for value in formatplus(returnData):
-        arr.append(value)
 
-    # Formating array with given keys using zip
-    formatarr = list(zip(xdarr, arr))
+# Global Arrays belong to rec
+arrkeys = ['Gender', 'Title', 'First Name', 'Last Name', 'Address', 'City', 'State', 'Country', 'Postal Code', 'Latitude', 'Longitude', 'Time Zone', 'Description', 'Email', 'UUID', 'User name', 'Password', 'Salt', 'MD5 Version', 'Sha1', 'Sha256', 'Date', 'Age', 'Date', 'Age', 'Phone Number', 'Cell Phone Number', 'Name', 'Value', 'Large', 'Medium', 'Thumbnail', 'NAT']
+arr = []
 
-    # Calling writedata function for save the file
-    writedata(formatarr)
+
+def rec(dicto):
+    if isinstance(dicto, dict) is False:
+        pass
+    keys = list(dicto.keys())
+
+    for k in keys:
+        if isinstance(dicto[k], (list, dict)):
+            rec(dicto[k])
+        else:
+            # only append the array of values
+            arr.append(str(dicto[k]))
+
+
+def bindstr(arrkeys, arr):
+    # returns a dictionary
+    # arguments are from global
+    major = dict(zip(arrkeys, arr))
+    return major
 
 
 def writedata(array):
-    # use person name as text file name
-    personName = array[0][1] + " " + array[1][1] + ".txt"
-    
-    # Creating text file using context manager
+    personName = str(array['First Name']) + "_" + str(array['Last Name'])
     with open(personName, 'w+') as file:
         file.write("THIS FILE IS GENERATED BY FAKE PERSON GENERATOR\n")
 
         # writing array information to text file using for loop
-        for key, value in array:
-            file.write(f'{key} : {value}\n')
-
-
-# Accessing JSON data using glom and add all returned data into a array
-def formatplus(returnData):
-
-    # Using glom to access data in the JSON object.
-    # Using simple generator to return the data 
-    yield glom(returnData, 'results.0.name.first')
-    yield glom(returnData, 'results.0.name.last')
-    yield glom(returnData, 'results.0.gender')
-    yield str(glom(returnData, 'results.0.location.street.number')) + " " + glom(returnData, 'results.0.location.street.name')
-    yield glom(returnData, 'results.0.location.city')
-    yield glom(returnData, 'results.0.location.state')
-    yield glom(returnData, 'results.0.location.country')
-    yield glom(returnData, 'results.0.location.coordinates.latitude')
-    yield glom(returnData, 'results.0.location.coordinates.longitude')
-    yield str(glom(returnData, 'results.0.location.postcode'))
-    yield glom(returnData, 'results.0.location.timezone.offset') + " " + glom(returnData, 'results.0.location.timezone.description')
-    yield glom(returnData, 'results.0.email')
-    yield glom(returnData, 'results.0.dob.date')
-    yield str(glom(returnData, 'results.0.dob.age'))
-    yield str(glom(returnData, 'results.0.phone'))
-    yield str(glom(returnData, 'results.0.cell'))
+        for key, value in array.items():
+            file.write(f'{key}  ::  {value}\n')
 
 
 # Created GUI Window with Tkinter
